requirements.txt
scapy>=2.5.0
pycryptodome>=3.19.0
cryptography>=41.0.0
psutil>=5.9.0
PyQt5>=5.15.0
pytest>=7.4.0
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.66.0
# netifaces yerine alternatif paketler
ifaddr>=0.2.0
netaddr>=0.8.0 